Move BaseEntity to adapters package.
